+++ experiments/billable/docs/Email.md
@@ -3,3 +3,4 @@
 Right now Cloudflare doesn't offer a solution for sending emails. Resend is probably the next best bet, but requires some setup.
 We will attempt to solve this for Reloaded users as soon as possible.
 
+^^ This is false. Get the example working.
