+++ experiments/billable/worker-configuration.d.ts
@@ -2,7 +2,9 @@
 
 interface Env {
 	SECRET_KEY: "secret";
-	SESSION_DO: DurableObjectNamespace<import("./dist/worker/worker").SessionDO>;
+	RESEND_API_KEY: "123";
+	APP_URL: "http://localhost:5173";
+	SESSION_DO: DurableObjectNamespace<import("./src/worker").SessionDO>;
 	R2: R2Bucket;
 	DB: D1Database;
 	ASSETS: Fetcher;
